Move normalizeconfig to separate file, passing configFile to make it pure
